Add index page
This adds a template for an index page, as well as a base template that
other pages can be derived from in the future. Since the templates are
processed by Jinja twice (once by cookiecutter, and once when the tool
is actually running), we have to mostly wrap them in raw blocks, though
a few parts are actually substituted at the cookiecutter level.