Add form with CSRF protection
This form on the “praise” endpoint allows you to customize the praise
you receive (yes, it’s silly). It demonstrates session manipulation, but
the main point is the CSRF protection. The way it’s set up here, the
CSRF token is tied to the session and remains valid until it is used
when a form is actually submitted; if you open several tabs with the
form, any one of them will be able to submit first, but afterwards all
the other ones will fail.

Right now, CSRF failure is silent (you just see the form again, with no
indication that anything went wrong), but I’ll improve on that soon.