Repeat form inputs on CSRF error
If the form failed to submit due to a CSRF protection error, fill in the
form with the user’s inputs so they can re-submit it (with a new CSRF
token) right away. (This should be easy to other cases where the form
has to be re-submitted for some reason.)

This implementation is not how I’ve done this in any of my other tools,
but I think it’s a fairly simple approach that should work well for most
simple forms. It uses more global state than I’d usually like, but it’s
global state that Flask seems to encourage (request context and
application context), so I figure it’s okay to go along with it.