Use toolforge library to set user agent
The library can set the default user agent for the requests library for
us, generated from the tool name and user email address (which we
default to the Toolforge email address for the tool, just as the
toolforge library does if you leave it unspecified). Later, we will also
need to set the user agent explicitly in some requests, so we also get
it out of the requests library again to store it in a variable.