Add OAuth authentication
This adds an authentication area with a login link or the currently
logged in username to the end of the navbar. The information isn’t used
otherwise yet, but that will come later.

If you access the OAuth endpoints and OAuth is not configured, the
request errors out, but that doesn’t affect subsequent requests (Flask
encapsulates them sufficiently), so while that might be fixed in the
future, I don’t see it as a priority.