Use OAuth on praise page
We use OAuth to get the name of the user to be praised, and to correctly
gender the default praise if none was explicitly set. (We could also get
the user name from the OAuth identity, as in authentication_area(), but
that doesn’t include the gender. Also, this demonstrates how to make
authenticated MediaWiki API requests, which is nice.)