Fix submitted_request_valid() for missing referrer
If a client does not send a Referer header, then flask.request.referrer
is None and we can’t call .startswith(…) on it. Fix this by falling back
to an empty string in that case, which is fine since the empty string
can never start with the index URL (which contains at least a slash), so
this case is correctly identified as a referrer mismatch.

This bug was identitified by mypy in the QuickCategories tool (see
lucaswerkmeister/tool-quickcategories@af3ec9b4e8).