Add option to set up pytest
This sets up some tests for csrf_token() (copied from the lexeme-forms
tool) and one for the praise endpoint.

The conditional file addition/removal is implemented as a post-generate
hook. On Unix platforms, directly enclosing the file name in {% if
cookiecutter.set_up_pytest %} would also work, but that would make the
template unusable on Windows, since % is not a valid file name character
there. I don’t know if this template currently works on Windows, but if
it does, I don’t want to directly jeopardize it like this.