Add option to set up mypy
This adds type hints to app.py, a basic stub for the toolforge library,
and “type: ignore” comments to some imports lacking type information.
Type checks can be run using `MYPYPATH=stubs/ mypy .`; this isn’t
documented anywhere in the template yet, but I plan to add a Makefile
for this (and also flake8 and pytest) next.

One quirk here is that cookiecutter doesn’t really support boolean
variables yet, so what we actually see in Jinja2 when processing the
app.py template is the strings "True"/"False", not the bools True/False,
which makes the conditions a bit less pretty.

The mypy version we pin in requirements.txt is the last one to support
Python 3.4, which we still need to support for now since Python 3.5 is
not yet available in Toolforge (though it hopefully will be soon).