Add option to set up Travis CI
This configures Travis to run the automated checks on Python 3.4
(current default version on Toolforge) and Python 3.5 (hopefully
available soon). If all the automated checks were disabled but Travis CI
was still enabled, we emit a warning but don’t abort the project
generation, in case the user wants to manually set up a Makefile (or
configure a different script in .travis.yml).