Add alternative, simpler CSRF protection
This uses the same token approach, but relieves the developer of having
to handle it in every POST, instead throwing a very lazy error on any
invalid POST. This is safer because you can’t forget the handling
anywhere, a lesson I learned in QuickCategories (from which this is more
or less copied). However, let’s still show developers how to handle CSRF
errors in each request handler if they want to.