Update for toolforge.org migration
With the toolforge.org migration [1], each tool gets a dedicated domain,
which means we can remove the referrer-based CSRF protection, as well as
the APPLICATION_ROOT config. We keep the anti-iframe protection because,
as Amir pointed out, it still protects tools from clickjacking.

This doesn’t yet update the tool to use Webservice templates [2], though
I hope to do that soon as well, because those are nice.

[1]: https://wikitech.wikimedia.org/wiki/News/Toolforge.org
[2]: https://wikitech.wikimedia.org/wiki/News/Toolforge.org#Webservice_templates