Keep CSRF tokens valid for longer
Invalidating the CSRF token on each form submit is inconvenient: it
prevents users from using the tool in multiple tabs at once, because the
different tabs will fight over the last valid CSRF token. It seems to be
acceptable to keep the token valid for longer (MediaWiki’s CSRF tokens
are also valid for a long time); to provide *some* way to invalidate it,
clear the token on each successful login.

test_app.py is updated to test the behavior with a *wrong* CSRF token
instead of an expired one (because the token no longer expires in the
tested scenario).