Allow continuing with invalid tool name
Generating a tool with an invalid name can be useful for several
reasons: to rewrite an existing tool, for throwaway testing, or in case
of Striker hiccups. Prompt the user instead of exiting unconditionally.
(However, aborting remains the default.)

Fixes #1.

Also, print only a short summary of the message in bold, not the whole
thing. That seems nicer to me.