Initialize Git repository in post_gen_project hook
Instead of telling the user to create a Git repository in the index
page, do it immediately after creating the project. (However, we don’t
go as far as creating the initial commit already.) Also, add an initial
.gitignore (I think I used to be hesitant to do that because the
template used to be VCS-agnostic, but that’s not the case anymore).

This respects the init.defaultBranch git config variable present in
recent Git versions, but if it’s not set, we pretend the default is
“main”, rather than “master” as it still is in current Git versions.
Also, use “main” in the index.html to refer to the branch.