Remove instruction to symlink service.template into ~
The new version of the `webservice` command automatically looks for this
file in various locations, including ~/www/python/src/, so symlinking
the file into the tool home directory is no longer necessary. [1]

(In fact, the command also complains if it finds more than one file,
even if one is just a symlink to the other, effectively breaking the
workflow of anyone who followed what I would describe as the previous
best practice. I don’t appreciate this, but I can’t be bothered to
advocate against it – with the new webservice version already rolled
out, the ship has sailed, I suppose.)

[1]: https://lists.wikimedia.org/hyperkitty/list/cloud-announce@lists.wikimedia.org/thread/U42VYJXEENMOTVWJZ2DTV4PNZXT6M7PY/