Use flask.config.from_file() to load config file
This is new in Flask 2.0.0 (previously, there was only from_json()), and
seems nicer than reading the file manually. Note that it only reads
all-uppercase keys (silently ignoring the rest), so the oauth key has to
be renamed to OAUTH; anyone backporting this change to their existing
tool will need to adjust existing config files.