Add types-PyYAML to requirements.txt
Apparently more recent Mypy versions don’t come with the whole typeshed
by default anymore, so we have to explicitly install some package if we
want types that don’t come with the main package.

Also, while we’re at it, adjust the capitalization of the PyYAML package
(it seems to be capitalized this way on PyPI).

The {% if %} / {% endif %} arrangement around the types-PyYAML line is
different than for the other lines, because otherwise, the file would
have to end with {% endif %} without a trailing newline in order to
avoid a double newline in the processed file, which is ugly.