Use `install -m` to create config file with correct mode
Copying the file and then chmod’ing it makes it theoretically possible
for an adversary to open(2) the file while it’s still world-readable and
then read(2) from it later, after it contains credentials: permissions
are only checked when opening the file, not at each read. Use the
`install` command to copy the file with the correct mode right away.
(Alternatively, we could adjust the umask, but install seems better in
this case.)

This also has the advantage of being shorter than the original two
commands :)