Remove matrix strategy from GitHub actions
I think the usefulness of this is limited: most tools can probably
migrate from one Python version to another in a single commit, and
replace the Python version in CI at the same time. Let’s simplify the
setup for that common case. (Part of the motivation is that matrix runs
are a bit more annoying to navigate on GitHub, even when there’s only
one instance of them.)