Revert "Don’t use mwoauth.identify() to get user name"
This was poorly advised, as index.php told users to use a “user
identification only“ consumer if they didn’t need any grants, but the
authorization area would be broken in that case: such consumers aren’t
allowed to use the API (error: readapidenied). I think the best solution
is to switch back to mwoauth.identify(): yes, it’s not a great
interface, but it’s better than only supporting “regular” OAuth
consumers in the tool, or than introducing another prompt in
cookiecutter.json asking which type of OAuth consumer users are
intending to use, or than including the code for both versions and
falling back from action='userinfo' to mwoauth.identify() if the former
yields an error.

This reverts commit a4b9d93e4bce6e68ab5b6540071807e6efeccb49.

Bug: T403932